removed fish from home
I've come to realize that good defaults are better. I can be more
efficient if I learn one tool and stick to it - not jumping from one
tool to the next unless there's a compelling reason to. Fish may be a
better shell. But it's not as universal as BASH which means I have to
learn both.